flytt validering spesifikt for barnets alder vilkår ut i egen fil
